requirements.txt
# Core dependencies
python-telegram-bot>=20.5
psycopg2-binary>=2.9.9
APScheduler>=3.10.4
python-dotenv>=1.0.0
alembic>=1.12.1

# Database
SQLAlchemy>=2.0.23

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Code quality
black>=23.11.0
flake8>=6.1.0
pylint>=3.0.3

# Optional: for calendar functionality
python-telegram-bot-calendar>=1.0.5

# Timezone support
pytz>=2024.1